feat: add analysis form
