+++ src/pages/diaries/new.tsx
@@ -2,7 +2,6 @@ import { useEffect } from "react";
 import { useRouter } from "next/router";
 
 import { useAuthContext } from "context/AuthContext";
-
 import DiaryForm from "components/entries/diaries/DiaryForm";
 
 export default function NewDiaryPage() {
